pyproject.toml
[project]
name = "ai-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.0",
    "langchain>=0.3.28",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.83",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=0.3.1",
    "onnxruntime>=1.24.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.55.0",
]
